🔧 Wrap Topbar elements in a link
The main elements in the topbar (including the Datahub logo and title) are now wrapped in a MudLink, thereby providing a direct way to navigate to the homepage by simply clicking on these elements. A small code formatting adjustment has also been made, placing the "SupportFormUrl" property declaration on one line.