🔧 Update TerraformOutput methods and add tests
This commit updates several methods in TerraformOutputHandler from private to internal. Moreover, TerraformOutputHelper's GetExpectedTerraformOutput method is refactored, extracting code blocks and the possibility to pass workspaceId and workspaceUrl as parameters. It also adds "Datahub.Functions.UnitTests" to visible internals in Datahub.Functions.csproj. Additionally, Unit tests are added to TerraforOutputHandlerTests to validate the processing of AzureStorageBlob and AzureDatabricks output variables.