🐛 Add "tool provisioning" status to the databricks card
A new label "Tool is being provisioned" was added to the localization files for both English and French languages. Additionally, a check status for tool provisioning was introduced in the Databricks card file. If the project resource is not created yet, the "Tool is being provisioned" message will be displayed along with a loading spinner, replacing the "Open Databricks" button which will be disabled. This improves user feedback during the tool provisioning process.