🔧 Update workspace validation to always use latest version
The previous behavior of the ValidateWorkspaceVersion method was to maintain the existing version if it was set to DefaultVersion. This has been modified so that it now always updates the Terraform workspace version to the latest version, regardless of the incoming version. This ensures that we're always using the most recent version.