Added email notifications when workspace resources deleted via inactivity (#1387)
* 🔧 Refactored TerraformStatus into its own class in the shared project

* 🧪 Add SpecFlow tests for Project Usage Notifier

Introduced tests to verify Project Usage Notifier functionality in scenarios where project usage exceeds or does not exceed the budget. Implemented required steps and hooks to simulate project data, and updated the ProjectUsageNotifier to handle over-budget scenarios effectively.

* 🧪 Add test for resource deletion when usage exceeds budget

Added a new scenario to the ProjectUsageNotifier feature file to test if a workspace's resources are set to deleted when usage surpasses the budget. Implemented the corresponding steps in ProjectUsageNotifierSteps.cs, including the verification logic and assertion.

* 🧪 Add tests for resource messaging in ProjectUsageNotifier

Extended tests to cover scenarios where resource messaging service should and should not be notified. Modified `ProjectUsageNotifier` to handle cases with no undeleted resources and included associated dependency injection in test hooks.

* 🧪 Prevent auto-delete for over-budget projects

Added logic to prevent auto-deletion for projects with the "prevent auto delete" flag set to true. Updated tests to cover this scenario and ensure resources are not marked for deletion in such cases.

* Update Datahub.Core.csproj

* Update Datahub.Core.csproj

* 📜 Update .gitignore to exclude SCSS files

Added a rule to ignore all files with the .scss extension. This helps to ensure that SCSS files are not tracked or committed to the repository. Also added a newline for better formatting.

* 🧪 Add conversion for Terraform template names to readable format

Implemented methods to convert template names to readable strings in English and French within `TerraformTemplate.cs`. Added related SpecFlow tests and feature definitions to ensure correct functionality and coverage.

* 🔧 Add email notifications for resource deletions

Implemented functionality to send email notifications to workspace administrators when resources are queued for deletion due to budget constraints. Added a new email template `delete_notification.html` and updated the project file to copy this template to the output directory.

* 🧪 Add email notification for overbudget workspace deletions

Implemented functionality to notify workspace leads and admin users via email when a workspace's resource is deleted due to overbudget conditions. Added corresponding tests to ensure correct behavior.

* Update delete_notification.html

fix french

---------

Co-authored-by: John Bain <john.bain@gmail.com>
Co-authored-by: Erik <erik.putrycz@ssc-spc.gc.ca>